// src/app/shared/types/article.interfase.ts
import {ProfileInterface} from './profile.interface';

export interface ArticleInterface {
  slug: string;
  title: string;
  description: string;
  body: string;
  tagList: string[];
  createdAt: string;
  updatedAt: string;
  favorited: boolean;
  favoritesCount: number;
  author: ProfileInterface;
}